now roles are finished. Tests pass.
